+++ portal-kernel/src/com/liferay/portal/kernel/portlet/LiferayPortletConfig.java
@@ -14,6 +14,8 @@
 
 package com.liferay.portal.kernel.portlet;
 
+import aQute.bnd.annotation.ProviderType;
+
 import com.liferay.portal.kernel.model.Portlet;
 
 import javax.portlet.PortletConfig;
@@ -22,6 +24,7 @@ import javax.portlet.PortletContext;
 /**
  * @author Brian Wing Shun Chan
  */
+@ProviderType
 public interface LiferayPortletConfig extends PortletConfig {
 
 	public static final String RUNTIME_OPTION_ESCAPE_XML =
